student/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.3.4=pyhd8ed1ab_0
brotlipy=0.7.0=py38h5406a74_1001
ca-certificates=2021.5.30=h033912b_0
certifi=2021.5.30=py38h50d1736_0
cffi=1.14.4=py38h6afc60c_0
chardet=4.0.0=py38h50d1736_1
configparser=5.0.2=pyhd3eb1b0_0
cryptography=3.4.7=py38h1fa4640_0
defusedxml=0.7.1=pyhd8ed1ab_0
diff-match-patch=20200713=pyh9f0ad1d_0
django=3.2=pyhd8ed1ab_0
django-cors-headers=3.7.0=pyh44b312d_0
django-import-export=2.5.0=pyhd8ed1ab_0
djangorestframework=3.12.4=pyhd8ed1ab_0
et_xmlfile=1.0.1=py_1001
idna=2.10=pyh9f0ad1d_0
jdcal=1.4.1=py_0
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libcxx=10.0.0=1
libedit=3.1.20210216=h9ed2024_1
libffi=3.2.1=h0a44026_1007
libgfortran=5.0.0=9_3_0_h6c81a4c_22
libgfortran5=9.3.0=h6c81a4c_22
liblapack=3.9.0=8_openblas
libopenblas=0.3.12=openmp_h54245bb_1
llvm-openmp=11.1.0=hda6cdc1_1
markuppy=1.14=py_0
ncurses=6.2=h0a44026_1
numpy=1.20.2=py38had91d27_0
odfpy=1.4.1=py_0
openpyxl=3.0.7=pyhd8ed1ab_0
openssl=1.1.1k=h0d85af4_0
pandas=1.0.1=py38h4f17bb1_0
pip=21.0.1=py38hecd8cb5_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py38h50d1736_3
python=3.8.0=h359304d_2
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py38h5406a74_0
readline=7.0=h1de35cc_5
requests=2.25.1=pyhd3deb0d_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=hffcf06c_0
sqlparse=0.4.1=pyh9f0ad1d_0
tablib=3.0.0=pyhd8ed1ab_0
tabulate=0.8.9=pyhd8ed1ab_0
tk=8.6.10=hb0a8c7a_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
xlrd=2.0.1=pyhd8ed1ab_3
xlwt=1.3.0=py_1
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h1de35cc_3
